+++ src/backends/sentry_backend_crashpad.cpp
@@ -253,6 +253,10 @@ flush_external_crash_report(
         sentry__envelope_add_session(envelope, options->session);
     }
 
+    if (options->cache_keep) {
+        sentry__envelope_set_header(envelope, "cache_dir",
+            sentry_value_new_string(options->run->cache_path->path));
+    }
     sentry__run_write_external(options->run, envelope);
     sentry_envelope_free(envelope);
 }
@@ -1012,7 +1016,8 @@ crashpad_backend_prune_database(sentry_backend_t *backend)
     crashpad::PruneCrashReportDatabase(data->db, &condition);
 }
 
-#if defined(SENTRY_PLATFORM_WINDOWS) || defined(SENTRY_PLATFORM_LINUX)
+#if defined(SENTRY_PLATFORM_WINDOWS) || defined(SENTRY_PLATFORM_LINUX)         \
+    || defined(SENTRY_PLATFORM_MACOS)
 static bool
 ensure_unique_path(sentry_attachment_t *attachment)
 {
@@ -1101,7 +1106,8 @@ sentry__backend_new(void)
     backend->user_consent_changed_func = crashpad_backend_user_consent_changed;
     backend->get_last_crash_func = crashpad_backend_last_crash;
     backend->prune_database_func = crashpad_backend_prune_database;
-#if defined(SENTRY_PLATFORM_WINDOWS) || defined(SENTRY_PLATFORM_LINUX)
+#if defined(SENTRY_PLATFORM_WINDOWS) || defined(SENTRY_PLATFORM_LINUX)         \
+    || defined(SENTRY_PLATFORM_MACOS)
     backend->add_attachment_func = crashpad_backend_add_attachment;
     backend->remove_attachment_func = crashpad_backend_remove_attachment;
 #endif