+++ src/backends/sentry_backend_inproc.c
@@ -16,6 +16,7 @@
 #endif
 #include "sentry_scope.h"
 #include "sentry_screenshot.h"
+#include "sentry_session_replay.h"
 #include "sentry_sync.h"
 #include "sentry_transport.h"
 #include "sentry_unix_pageallocator.h"
@@ -1158,6 +1159,17 @@ process_ucontext_deferred(const sentry_ucontext_t *uctx,
                     sentry__envelope_add_attachment(envelope, screenshot);
                 }
                 sentry__attachment_free(screenshot);
+            }
+
+            if (options->attach_session_replay) {
+                sentry_attachment_t *replay = sentry__attachment_from_path(
+                    sentry__session_replay_get_path(options));
+                if (replay
+                    && sentry__session_replay_capture(
+                        replay->path, options->session_replay_duration, 0)) {
+                    sentry__envelope_add_attachment(envelope, replay);
+                }
+                sentry__attachment_free(replay);
             }
 
             if (!sentry__launch_external_crash_reporter(options, envelope)) {