+++ tests/test_integration_crashpad.py
@@ -285,8 +285,8 @@ def wait_for_no_werfault(timeout=30.0, poll_interval=0.5):
     reason="Test covers Windows-specific crashes which can only be covered via the Crashpad WER module",
 )
 # this test currently can't run on CI because the Windows-image doesn't properly support WER, if you want to run the
-# test locally, invoke pytest with the --with_crashpad_wer option which is matched with this marker in the runtest setup
-@pytest.mark.with_crashpad_wer
+# test locally, invoke pytest with the --with_wer option which is matched with this marker in the runtest setup
+@pytest.mark.with_wer
 @pytest.mark.parametrize(
     "run_args",
     [
@@ -369,18 +369,10 @@ def test_crashpad_wer_crash(cmake, httpserver, run_args):
         pytest.param(
             ["attach-after-init"],
             {},
-            marks=pytest.mark.skipif(
-                sys.platform == "darwin",
-                reason="crashpad doesn't support dynamic attachments on macOS",
-            ),
         ),
         pytest.param(
             ["attachment", "attach-after-init", "clear-attachments"],
             {},
-            marks=pytest.mark.skipif(
-                sys.platform == "darwin",
-                reason="crashpad doesn't support dynamic attachments on macOS",
-            ),
         ),
     ],
 )
@@ -768,6 +760,7 @@ def test_crashpad_external_crash_reporter(cmake, httpserver, run_args):
     tmp_path = cmake(
         ["sentry_example", "sentry_crash_reporter"], {"SENTRY_BACKEND": "crashpad"}
     )
+    cache_dir = tmp_path.joinpath(".sentry-native/cache")
 
     env = dict(os.environ, SENTRY_DSN=make_dsn(httpserver))
     httpserver.expect_oneshot_request("/api/123456/envelope/").respond_with_data("OK")
@@ -777,7 +770,7 @@ def test_crashpad_external_crash_reporter(cmake, httpserver, run_args):
         run(
             tmp_path,
             "sentry_example",
-            ["log", "crash-reporter"] + run_args,
+            ["log", "crash-reporter", "cache-keep"] + run_args,
             expect_failure=True,
             env=env,
         )
@@ -791,6 +784,7 @@ def test_crashpad_external_crash_reporter(cmake, httpserver, run_args):
     crash = crash_request.get_data()
 
     envelope = Envelope.deserialize(crash)
+    assert envelope.headers["cache_dir"] == str(cache_dir)
     assert_meta(envelope, integration="crashpad")
     assert_breadcrumb(envelope)
 
@@ -803,8 +797,8 @@ def test_crashpad_external_crash_reporter(cmake, httpserver, run_args):
     reason="Test covers Windows-specific crashes which can only be covered via the Crashpad WER module",
 )
 # this test currently can't run on CI because the Windows-image doesn't properly support WER, if you want to run the
-# test locally, invoke pytest with the --with_crashpad_wer option which is matched with this marker in the runtest setup
-@pytest.mark.with_crashpad_wer
+# test locally, invoke pytest with the --with_wer option which is matched with this marker in the runtest setup
+@pytest.mark.with_wer
 @pytest.mark.parametrize(
     "run_args",
     [
@@ -815,8 +809,15 @@ def test_crashpad_external_crash_reporter_wer(cmake, httpserver, run_args):
     test_crashpad_external_crash_reporter(cmake, httpserver, run_args)
 
 
-@pytest.mark.parametrize("cache_keep", [True, False])
-def test_crashpad_cache_keep(cmake, httpserver, cache_keep):
+@pytest.mark.parametrize(
+    "cache_args,expect_cache",
+    [
+        ([], False),
+        (["cache-keep"], True),
+        (["cache-keep-always"], True),
+    ],
+)
+def test_crashpad_cache_keep(cmake, httpserver, cache_args, expect_cache):
     tmp_path = cmake(["sentry_example"], {"SENTRY_BACKEND": "crashpad"})
     cache_dir = tmp_path.joinpath(".sentry-native/cache")
 
@@ -827,7 +828,7 @@ def test_crashpad_cache_keep(cmake, httpserver, cache_keep):
         run(
             tmp_path,
             "sentry_example",
-            ["log", "crash"] + (["cache-keep"] if cache_keep else []),
+            ["log", "crash"] + cache_args,
             expect_failure=True,
             env=env,
         )
@@ -839,7 +840,7 @@ def test_crashpad_cache_keep(cmake, httpserver, cache_keep):
     run(
         tmp_path,
         "sentry_example",
-        ["log", "no-setup"] + (["cache-keep"] if cache_keep else []),
+        ["log", "no-setup"] + cache_args,
         env=env,
     )
 
@@ -847,12 +848,12 @@ def test_crashpad_cache_keep(cmake, httpserver, cache_keep):
     run(
         tmp_path,
         "sentry_example",
-        ["log", "no-setup"] + (["cache-keep"] if cache_keep else []),
+        ["log", "no-setup"] + cache_args,
         env=env,
     )
 
-    assert cache_dir.exists() or cache_keep is False
-    if cache_keep:
+    assert cache_dir.exists() or expect_cache is False
+    if expect_cache:
         cache_files = list(cache_dir.glob("*.envelope"))
         assert len(cache_files) == 1
         with open(cache_files[0], "rb") as f:
@@ -864,6 +865,8 @@ def test_crashpad_cache_keep(cmake, httpserver, cache_keep):
         dmp_files = list(cache_dir.glob("*.dmp"))
         assert len(dmp_files) == 1
         assert cache_files[0].stem == dmp_files[0].stem
+    else:
+        assert not cache_dir.exists() or len(list(cache_dir.glob("*.envelope"))) == 0
 
 
 def test_crashpad_cache_consent(cmake, httpserver):